uos_grants/requirements.txt
pydantic
langchain-ollama
langchain-text-splitters
langchain-community
langgraph
langchain-openai
langchain-core
langchain-pinecone
beautifulsoup4
langchain-google-genai
asyncpg
langchain_postgres
psycopg[binary,pool]
